# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is our project server   ",
    "version": "1.0.0",
    "title": "Swagger project",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost",
  "basePath": ":3000/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/qeustion": {
      "post": {
        "tags": [
          "question"
        ],
        "summary": "Create question",
        "description": "",
        "operationId": "questionid",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created question object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Question"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
      
    
      
    },
...
    },
  "securityDefinitions": {
    "app_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://project.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:users": "modify users ",
        "read:users": "read your users"
      }
    }
  
  },
  "definitions": {
...
    "Question": {
      "type": "object",
      "properties": {
        "questionid": {
          "type": "integer",
          "format": "int64"
        },
        "userid": {
          "type": "integer",

          "format": "int64"
        },
       
        "content": {
          "type": "string"
         
        }
       
      },
      "xml": {
        "name": "Question"
      }
    },
...
  },
...
}